# openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "FastAPI", "version": "0.1.0" },
  "paths": {
...
    "/v1/customers": {
      "get": {
        "summary": "List Customers",
        "operationId": "list_customers_v1_customers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CustomerList" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Customer",
        "operationId": "create_customer_v1_customers_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_customer_v1_customers_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Customer" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/v1/customers/{customer_id}": {
      "post": {
        "summary": "Update Customer",
        "operationId": "update_customer_v1_customers__customer_id__post",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Customer Id" }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Body_update_customer_v1_customers__customer_id__post"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Customer" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Retrieve Customer",
        "operationId": "retrieve_customer_v1_customers__customer_id__get",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Customer Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    { "$ref": "#/components/schemas/Customer" },
                    { "type": "null" }
                  ],
                  "title": "Response Retrieve Customer V1 Customers  Customer Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Customer",
        "operationId": "delete_customer_v1_customers__customer_id__delete",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Customer Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CustomerDelete" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/v1/customers/search": {
      "get": {
        "summary": "Search Customers",
        "operationId": "search_customers_v1_customers_search_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CustomerList" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/v1/customers/{customer_id}/addresses": {
      "post": {
        "summary": "Create Address",
        "operationId": "create_address_v1_customers__customer_id__addresses_post",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Customer Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_address_v1_customers__customer_id__addresses_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CustomerAddress" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Customer Addresses",
        "operationId": "list_customer_addresses_v1_customers__customer_id__addresses_get",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Customer Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CustomerAddressList" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/v1/customers/{customer_id}/addresses/{address_id}": {
      "post": {
        "summary": "Update Address",
        "operationId": "update_address_v1_customers__customer_id__addresses__address_id__post",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Customer Id" }
          },
          {
            "name": "address_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Address Id" }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Body_update_address_v1_customers__customer_id__addresses__address_id__post"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CustomerAddress" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Retrieve Customer Address",
        "operationId": "retrieve_customer_address_v1_customers__customer_id__addresses__address_id__get",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Customer Id" }
          },
          {
            "name": "address_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Address Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CustomerAddress" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Customer Address",
        "operationId": "delete_customer_address_v1_customers__customer_id__addresses__address_id__delete",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Customer Id" }
          },
          {
            "name": "address_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Address Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerAddressDelete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/v1/options": {
      "get": {
        "summary": "List Options",
        "operationId": "list_options_v1_options_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OptionList" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Option",
        "operationId": "create_option_v1_options_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_option_v1_options_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Option" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/v1/options/{option_id}": {
      "post": {
        "summary": "Update Option",
        "operationId": "update_option_v1_options__option_id__post",
        "parameters": [
          {
            "name": "option_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Option Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_update_option_v1_options__option_id__post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Option" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Retrieve Option",
        "operationId": "retrieve_option_v1_options__option_id__get",
        "parameters": [
          {
            "name": "option_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Option Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Option" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Option",
        "operationId": "delete_option_v1_options__option_id__delete",
        "parameters": [
          {
            "name": "option_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Option Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OptionDelete" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_create_address_v1_customers__customer_id__addresses_post": {
        "properties": {
          "name": { "type": "string", "title": "Name" },
          "line1": { "type": "string", "title": "Line1" },
          "city": { "type": "string", "title": "City" },
          "state": { "type": "string", "title": "State" },
          "country": { "type": "string", "title": "Country" },
          "postal_code": { "type": "string", "title": "Postal Code" },
          "line2": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Line2"
          }
        },
        "type": "object",
        "required": [
          "name",
          "line1",
          "city",
          "state",
          "country",
          "postal_code"
        ],
        "title": "Body_create_address_v1_customers__customer_id__addresses_post"
      },
...
      "Body_create_customer_v1_customers_post": {
        "properties": {
          "name": { "type": "string", "title": "Name" },
          "phone": { "type": "string", "title": "Phone" },
          "email": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Email"
          },
          "address_name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Address Name"
          },
          "address_line1": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Address Line1"
          },
          "address_line2": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Address Line2"
          },
          "address_city": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Address City"
          },
          "address_state": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Address State"
          },
          "address_country": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Address Country"
          },
          "address_postal_code": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Address Postal Code"
          }
        },
        "type": "object",
        "required": ["name", "phone"],
        "title": "Body_create_customer_v1_customers_post"
      },
...
      "Body_create_option_v1_options_post": {
        "properties": {
          "name": { "type": "string", "title": "Name" },
          "values": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Values"
          },
          "product": { "type": "string", "title": "Product" }
        },
        "type": "object",
        "required": ["name", "values", "product"],
        "title": "Body_create_option_v1_options_post"
      },
...
      "Body_update_address_v1_customers__customer_id__addresses__address_id__post": {
        "properties": {
          "name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Name"
          },
          "line1": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Line1"
          },
          "line2": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Line2"
          },
          "city": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "City"
          },
          "state": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "State"
          },
          "country": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Country"
          },
          "postal_code": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Postal Code"
          }
        },
        "type": "object",
        "title": "Body_update_address_v1_customers__customer_id__addresses__address_id__post"
      },
...
      "Body_update_customer_v1_customers__customer_id__post": {
        "properties": {
          "name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Name"
          },
          "phone": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Phone"
          },
          "email": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Email"
          }
        },
        "type": "object",
        "title": "Body_update_customer_v1_customers__customer_id__post"
      },
...
      "Body_update_option_v1_options__option_id__post": {
        "properties": {
          "name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Name"
          },
          "values": {
            "anyOf": [
              { "items": { "type": "string" }, "type": "array" },
              { "type": "null" }
            ],
            "title": "Values"
          }
        },
        "type": "object",
        "required": ["name", "values"],
        "title": "Body_update_option_v1_options__option_id__post"
      },
...
      "Customer": {
        "properties": {
          "created": { "type": "integer", "title": "Created" },
          "updated": { "type": "integer", "title": "Updated" },
          "livemode": { "type": "boolean", "title": "Livemode" },
          "name": { "type": "string", "title": "Name" },
          "email": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Email"
          },
          "phone": { "type": "string", "title": "Phone" },
          "id": { "type": "string", "title": "Id" },
          "object": {
            "type": "string",
            "title": "Object",
            "default": "customer"
          },
          "addresses": {
            "anyOf": [
              { "$ref": "#/components/schemas/CustomerAddressList" },
              { "type": "null" }
            ]
          }
        },
        "type": "object",
        "required": ["created", "updated", "livemode", "name", "phone", "id"],
        "title": "Customer"
      },
      "CustomerAddress": {
        "properties": {
          "created": { "type": "integer", "title": "Created" },
          "updated": { "type": "integer", "title": "Updated" },
          "livemode": { "type": "boolean", "title": "Livemode" },
          "name": { "type": "string", "title": "Name" },
          "line1": { "type": "string", "title": "Line1" },
          "line2": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Line2"
          },
          "city": { "type": "string", "title": "City" },
          "state": { "type": "string", "title": "State" },
          "country": { "type": "string", "title": "Country" },
          "postal_code": { "type": "string", "title": "Postal Code" },
          "id": { "type": "string", "title": "Id" },
          "object": {
            "type": "string",
            "title": "Object",
            "default": "customer_address"
          }
        },
        "type": "object",
        "required": [
          "created",
          "updated",
          "livemode",
          "name",
          "line1",
          "city",
          "state",
          "country",
          "postal_code",
          "id"
        ],
        "title": "CustomerAddress"
      },
      "CustomerAddressDelete": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "object": {
            "type": "string",
            "title": "Object",
            "default": "customer_address"
          },
          "deleted": { "type": "boolean", "title": "Deleted" }
        },
        "type": "object",
        "required": ["id", "deleted"],
        "title": "CustomerAddressDelete"
      },
      "CustomerAddressList": {
        "properties": {
          "object": { "type": "string", "title": "Object", "default": "list" },
          "url": { "type": "string", "title": "Url" },
          "has_more": { "type": "boolean", "title": "Has More" },
          "data": {
            "items": { "$ref": "#/components/schemas/CustomerAddress" },
            "type": "array",
            "title": "Data",
            "default": []
          }
        },
        "type": "object",
        "required": ["url", "has_more"],
        "title": "CustomerAddressList"
      },
      "CustomerDelete": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "object": {
            "type": "string",
            "title": "Object",
            "default": "customer"
          },
          "deleted": { "type": "boolean", "title": "Deleted" }
        },
        "type": "object",
        "required": ["id", "deleted"],
        "title": "CustomerDelete"
      },
      "CustomerList": {
        "properties": {
          "object": { "type": "string", "title": "Object", "default": "list" },
          "url": {
            "type": "string",
            "title": "Url",
            "default": "/v1/customers"
          },
          "has_more": { "type": "boolean", "title": "Has More" },
          "data": {
            "items": { "$ref": "#/components/schemas/Customer" },
            "type": "array",
            "title": "Data",
            "default": []
          }
        },
        "type": "object",
        "required": ["has_more"],
        "title": "CustomerList"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Option": {
        "properties": {
          "created": { "type": "integer", "title": "Created" },
          "updated": { "type": "integer", "title": "Updated" },
          "livemode": { "type": "boolean", "title": "Livemode" },
          "name": { "type": "string", "title": "Name" },
          "values": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Values"
          },
          "id": { "type": "string", "title": "Id" },
          "object": { "type": "string", "title": "Object", "default": "option" }
        },
        "type": "object",
        "required": ["created", "updated", "livemode", "name", "values", "id"],
        "title": "Option"
      },
      "OptionDelete": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "object": {
            "type": "string",
            "title": "Object",
            "default": "option"
          },
          "deleted": { "type": "boolean", "title": "Deleted" }
        },
        "type": "object",
        "required": ["id", "deleted"],
        "title": "OptionDelete"
      },
      "OptionList": {
        "properties": {
          "object": { "type": "string", "title": "Object", "default": "list" },
          "url": { "type": "string", "title": "Url" },
          "has_more": { "type": "boolean", "title": "Has More" },
          "data": {
            "items": { "$ref": "#/components/schemas/Option" },
            "type": "array",
            "title": "Data",
            "default": []
          }
        },
        "type": "object",
        "required": ["url", "has_more"],
        "title": "OptionList"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}